Use tags = NULL in middle tables if object doesn't have any tags
This doesn't make much of a difference for the ways and rels table, but
if we store all nodes in the database, it does make a huge difference,
because most nodes don't have any tags. For a current planet, disk usage
for the nodes table goes from 476 GB down to 409 GB saving 67 GB or
nearly 15%.

Additionally it makes use of that table simpler. If you want to do any
queries on tags, you need an index on the tags column on the
nodes/ways/rels tables like this:

CREATE INDEX ON planet_osm_ways USING gin (tags);

But that is wasteful, because of the empty tags. We probably want to
generate them as

CREATE INDEX ON planet_osm_ways USING gin (tags) WHERE tags != '{}'::jsonb;

But now all queries on those tables have to include that extra condition
so that the query planner will use the index.

SELECT * FROM planet_osm_ways WHERE tags ? 'highway' AND tags != '{}'::jsonb;

If we use NULLs, the index can be created as:

CREATE INDEX ON planet_osm_ways USING gin (tags) WHERE tags IS NOT NULL;

And now the query becomes simpler, because the NOT NULL is automatically
taken into account by the query planner:

SELECT * FROM planet_osm_ways WHERE tags ? 'highway';

Note that this is an incompatible change to the new format middle
tables, but they are still marked as experimental, so we can do this.