Allow NULL values in new middle format members list
This makes osm2pgsql a bit more future proof by allowing the list of
members (which is encoded as JSON in the new middle format) to be empty,
i.e. to contain NULL. We currently don't write empty member lists as
NULL but as an empty JSON list, but if we change this in the future,
older versions of osm2pgsql will be able to read this correctly.