Refactor: Move storage of db connections from flex tables to output
Each table of the flex output kept its own database connection in the
table_connection_t class. With this change, the connection is kept
in a vector in the flex output instead.

This is a refactoring step towards using fewer database connections. As
long as each table had its own connection, we can not reduce them, but
if we keep them outside, we will be able to re-use the same connection
for different tables.