Disable synchronous commit on all connections
This simplifies the code, because we don't have to decide whether to set
this or not on each connection. And it doesn't matter much when setting
this for read-only connections which don't need this.

Setting synchronous_commit = off speeds up writing to the database,
because commis are not delayed until the WAL is written to disk. It
can not lead to corruption of the database.

https://www.postgresql.org/docs/current/runtime-config-wal.html#GUC-SYNCHRONOUS-COMMIT