Optionally build id index as unique index
Specify with `create_index = 'unique'` in the `ids` field of the
`define_table()` Lua function.

The user has to make sure not to insert multiple rows into such a table.

This can be useful if some program using the data has special handling
for unique ids/primary keys, for instance when it needs unique ids to
allow editing.