Generalization: Get extent of source layer also for raster layers
When generalizations are calculated by tile we need to know which extent
to work on. If we are not working with a full planet we get that
information from the (estimated) extent of the source layer. This adds
support for determining the extent for raster-based source layers.

This is needed when we generalize source layers into rasters and then
use those rasters as source for even lower level tiles.