New nodes and ways can't have existing parents
New nodes or ways (ie with version=1) can't have parent ways or
relations. So we don't have to add them to the list of "changed nodes"
or "changed ways" to trigger re-processing of changed parents. This will
save us a whole lot of effort looking for parents that are never there.

Note that we are checking for version != 1. In the (unlikely) case that
there are no versions in the input file (making version == 0),
everything will work as before.