Allow a limit to be set on expiry of line segments
Accidentally or deliberately moving a node a long way creates a
long line segment which can expire large numbers of tiles causing
disruption to the rendering process as osm2pgsql either runs out
of memory or generates a huge list of tiles to expire.

Limiting expiry of long line segments prevents that and also helps
reduce disruption of the resulting rendered maps with long lines
that are almost certainly incorrect.