Remove special case for old PostGIS versions when clustering
Ordering by geometry was done differently before PostGIS 2.4.
This commit removes the old code and adds a check that PostGIS 2.5
is available which is the oldest version we are checking in our CI.

Note that PostGIS 3 has been available for a long time now (it is in
Debian Bullseye ("oldstable", released 2021) and Ubuntu 20.04), so
we are way beyond where this should matter in real setups.