Allow empty config file in flex output
No flex config file, or one that doesn't define any tables is usually
not very useful. But it can be used in slim mode to have the middle
tables only. Or it can be used to do some other processing or instead
if the null output for testing.

This change allows an empty config file, but a warning is printed that
no output tables are defined. You still have to specify a config file,
it can be an empty one, but it has to be there. This is so that if the
user forgets naming a config file on the command line they get an error
and not only a warning that might be missed.