Add get_bbox() function to geometries in Lua
Add new get_bbox() function for geometries in Lua. Like the get_bbox()
function on OSM objects it returns the bounding box as four numbers in a
Lua array (table). Having this function on geometries allows for more
flexibility, you can call it on a centroid or a projected geometry for
instance.

It is faster to call get_bbox() on the OSM object, because it removes
the step of building the geometry first, so for those cases where this
is good enough the old use is fine and is still supported.

See #2219