Github actions: Add clang-tidy run to CI
Currently CI will not fail if there are warnings from clang-tidy. This
is, for the time being, only a way to easily see the messages.

Clang-tidy has its own workflow, so it doesn't interfere with the rest.
But because it also needs cmake to run, the cmake run has been taken
out of the 'build-and-test' action and put into its own.

This commit also disables some clang-tidy warnings we get from newer
versions of clang-tidy.