Make attributes of OSM objects always available to Lua
I did some benchmarking and it turns out there is real no measurable
difference between putting the data in there and not doing it. So
we'll just always do it. This means that it doesn't depend on the
-x/--extra-attributes command line option any more.

Note that the attributes will still be nil when coming from the
middle in update mode if -x/--extra-attributes was not used on import.