Remove support for non-bucket index on middle way table
Only supports bucket index which has been the default for a long time
and works much better than the old one.

Also removes the command line option --middle-way-node-index-id-shift.
The id shift can not be changed any more, it is hardcoded to 5 which
was the default.

This is a potentially breaking change: Users with incompatible indexes
have to do a reimport (or at least create the new index).