Simplify property handling and do updates all at once
Property changes were reflected piece by piece in the database. This
changes it so they are only done in one go after all the changes have
been made. So if something fails before we get to the point where we
write all properties to the database, no updates are actually done.

And the code is simpler because we keep track of necessary database
updates automatically and do them the next time the store() function is
called.