Store properties in database only later after more checks
If osm2pgsql fails during initialization, for instance because the input
file or the style file was not found, it had already initialized the
osm2pgsql_properties table in the database. We now write the properties
table later to reduce the chances that a small error on the command line
destroys an existing database this way.