Add support for node processing in second stage
This is completely analog to how way processing happens in the second
stage. So from the select_relation_member() function in the Lua config
file you can now also return a list of nodes that you request to process
again.

Note that this will only re-process the member nodes themselves, ways
which have these nodes as members are not re-processed. Also there is no
way to mark member nodes of ways, only member nodes of relations. So
this will allow processing, say, stop positions in public transport route
relations but not, say, barriers on roads.

This change is now possible because we removed support for the old
middle format which didn't allow storing the complete nodes (with tags).