Use protocol level prepare instead of PREPARE sql commands
Introduce a new pg_conn_t::prepare() function which is now used in
several places instead of pg_conn_t::exec() with a SQL PREPARE command.

This does not yet replace all places where PREPARE is used, the rest
will come in a later commit.

This is to make osm2pgsql work with some connection poolers that have
problems with prepared statements. For some background see
https://github.com/osm2pgsql-dev/osm2pgsql/discussions/2118 and
https://www.crunchydata.com/blog/prepared-statements-in-transaction-mode-for-pgbouncer