Check flat node file format for forwards compatibility
In a flat node file the first location is that of node id 0, which does
not exist. So the first location must always be the undefined location,
otherwise we might be looking at a different kind of file.

This adds a check for forwards compatibility. If and when we change the
file format, we can use the first 8 bytes to differentiate the file
formats.