Fix two-stage processing when flat node file is used
This is an additional fix to the fix in 8d1a55e5. If a flat node file is
used, nodes are not stored in the database (unless --middle-with-nodes
is specified, but also then only nodes with tags are stored in the
database).

There are two problems here:

1. The "get_node" prepared statement was not prepared in this case,
   so it fails when executes. This is fixed by only executing that
   statement if m_store_options.nodes is set in the middle.
2. If there is no node in the database, there might still be one
   in the flat node file. So we need to try this and create the
   node from the information in there. (It will only have the node
   id and location in it, but at least we get the location.)

This commit contains both fixes.