New feature: Check location of OSM objects against list of regions
This commit introduces a new feature: Locators. A locator is initialized
with one or more regions, each region has a name and a polygon or
bounding box. A geometry of an OSM object can then be checked against
this region list to figure out in which region(s) it is located.
This check is much faster than it would be to do this inside the
database after import.

Locators can be used for all sorts of interesting features:
* Read larger OSM file but import only data inside some area.
* Annotate each OSM object with the country (or other region) it
  is in. This can then, for instance, be used to show special
  highway shields for each country.
* Use the information which region the data is in for further
  processing, for instance setting of default values for the
  speed limit or using special language transliterations rules
  based on country.

Locators are created in Lua with `define_locator()`. Bounding boxes can
be added with `add_bbox()`. Polygons can be added from the database by
calling `add_from_db()` and specifiying an SQL query which can return
any number of rows each defining a region with the name and the
(multi)polygon as columns.

A locator can then be queried using `all_intersecting()` returning a
list of names of all regions that intersect the specified OSM object
geometry. Or the `first_intersecting()` function can be used which only
returns a single region for those cases where there can be no
overlapping data or where the details of objects straddling region
boundaries don't matter.

Several example config files are provided in the flex-config/locator
directory.