Fix two bugs in wildcard match
Two or more consecutive * characters were not handled. A fix was on the
web site this code comes from.

Before calling wildMatch recursively with "str + 1", the code did not
check whether *str was zero, in which case the "str + 1" is illegal.

This also changes the names of the function parameters which were not
consistent in .hpp and .cpp files.