Simplify and fix code getting relation members
This changes the way we are getting node members of relations. We now
always get the locations of those nodes directly in the middle code
instead of only getting the nodes themselves and adding the locations
through an extra call from the output.

This also fixes a problem where node locations for untagged nodes were
not retrieved from the RAM middle. The node_get() code does that
correctly, while the old code didn't.