Process untagged ways/relations in stage 1b/c correctly
When we introduced the "untagged" callbacks in
bc87ea2c8b4ea0bcac09975627e78765fda8927e, we did not call the
process_untagged_way/relation callbacks for untagged ways/relations  in
stage 1b and 1c, but called the normal process_way/relation callbacks.
That's not correct. So this is changed in this commit.

For stage2 processing the answer which callback to call is a bit more
complicated: On first glance we should also call the "untagged" function
for untagged objects. But if we call the "untagged" function in that
case, we would have to implement that function and it would be called
for the millions of untagged objects we are not interested in. So it is
arguably better to call the normal processing functions here. After all
we explicitly requested the processing function to be called with the
select_relation_members() call. So we are keeping this behaviour.