Fix: Always disable expiry if maxzoom==0
maxzoom==0 is the default setting and it is documented to means that
there is no expire. That's what the enabled() function checks.

But disabling this in from_bbox() only disabled the expire for that
case, not for linestrings. Better to do this check in the function
that's called from outside the expire code:  from_geometry().

We also need to fix some tests that were using maxzoom==0 by changing
their setting to use maxzoom==1.