Do area expire based on the actual polygon instead of the bbox
Expire (multi)polygons based on their shape, not the complete bounding
box. We use the existing implementation for the boundary expire, to get
the boundary right including the configured buffer. For the rest of the
area inside the polygon we use a "polygon drawing algorithm" to figure
out all tiles which are inside the polygon.

This also contains some changes to the BDD test code to allow running
the osm2pgsql-expire program and check its output against a file.