Flex: Allow creating point geometry from way
This allows the as_point() function to be called for ways. The function
has a single (optional) parameter specifying which member node to create
the point from. The index parameter is 1-based as usual in Lua. Negative
indexes count from the back,

:as_point(), :as_point(nil), :as_point(1) will return the first node,
:as_point(2) the second node and so on. :as_point(-1) will return the
last node.

If the index is 0 or out of range a null geometry will be returned. If
the parameter is not a number an error will be generated.