Reorganize expire code
The code was a bit weird, because the decision whether to do full area
expire of boundary only is mixed up with the code to do the full area
expire. This refactoring fixes that, we decide first what to do and then
do it.

What is a bit awkward now is the way we do the box calculation in the
decide_expire_mode() function only if needed and then later if it wasn't
done already. But if all goes well this will be cleaned up in a future
commit when we don't do expire by bounding box any more but by the
actual area.